A dream about trigger shapes: a tool to locate gap sequence based on logic analyzer file brute force #7947 only:progress
