+++ firmware/controllers/engine_cycle/fuel_schedule.cpp
@@ -114,7 +114,7 @@ expected<float> InjectionEvent::computeInjectionAngle() const {
 	wrapAngle(openingAngle, "addFuel#2", ObdCode::CUSTOM_ERR_6555);
 
 #if EFI_UNIT_TEST
-	printf("registerInjectionEvent openingAngle=%.2f inj %d\r\n", openingAngle, cylinderNumber);
+//	printf("registerInjectionEvent openingAngle=%.2f inj %d\r\n", openingAngle, cylinderNumber);
 #endif
 
 	return openingAngle;
