+++ unit_tests/tests/trigger/test_trigger_decoder_2.cpp
@@ -3,25 +3,7 @@
 using ::testing::_;
 using ::testing::StrictMock;
 
-class MockTriggerConfiguration : public TriggerConfiguration {
-public:
-	MockTriggerConfiguration(trigger_config_s type)
-		: TriggerConfiguration("Mock")
-		, m_type(type)
-	{ }
-
-protected:
-	bool isVerboseTriggerSynchDetails() const override {
-		return false;
-	}
-
-	trigger_config_s getType() const override {
-		return m_type;
-	}
-
-private:
-	const trigger_config_s m_type;
-};
+#include "mock_trigger_configuration.h"
 
 struct MockTriggerDecoder : public TriggerDecoderBase {
 	MockTriggerDecoder() : TriggerDecoderBase("mock") { }
